+++ e2e/hooks/javascript-hooks.e2e.js
@@ -66,8 +66,7 @@ describe('javascript-hooks', function () {
       expect(result.trim()).to.equal('got is-type and got is-string');
     });
     it('should be able to require the internal file using require(bit/) syntax', () => {
-      const appJsFixture =
-        "const isType = require('bit/utils/is-string/utils/internals/is-type'); console.log(isType());";
+      const appJsFixture = "const isType = require('bit/utils/is-string/internals/is-type'); console.log(isType());";
       fs.outputFileSync(path.join(helper.localScopePath, 'app.js'), appJsFixture);
       const result = helper.runCmd('node app.js');
       expect(result.trim()).to.equal('got is-type');