+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- avoid writing long files paths for imported components when there is a shared directory among the component files and its dependencies
+
 ## [0.11.1] - 2017-11-29
 
 - support tagging the entire local scope and all imported components to a specific tag using `--scope` and `--include_imported` flags